+++ test.py
@@ -25,7 +25,8 @@ init_term = m.parseTerm(init)
 init_term.reduce()
 n0 = Node(m, init_term)
 
-V0 = lambda node: 0
+#V0 = lambda node: 0
+V0 = lambda obs_term, g_state=None: 0
 
 print('\n=== SEARCH WITHOUT TRAINING ===')
 start_time = time.perf_counter()
